+++ ts/packages/knowledgeProcessor/src/conversation/knowledge.ts
@@ -122,10 +122,6 @@ export function createKnowledgeExtractorSettings(
     };
 }
 
-export interface ExtractedValue<TSourceId = any> {
-    sourceIds?: TSourceId[] | undefined;
-}
-
 export type ExtractedEntity<TSourceId = any> = {
     value: ConcreteEntity;
     sourceIds: TSourceId[];
